# tutor-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "家教管理后台系统接口文档",
        "title": "家教管理系统 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/exam-results": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "成绩管理"
                ],
                "summary": "获取所有成绩",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "学生ID",
                        "name": "student_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "course_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "考试类型",
                        "name": "exam_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ExamResult"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "成绩管理"
                ],
                "summary": "创建成绩记录",
                "parameters": [
                    {
                        "description": "成绩信息",
                        "name": "result",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ExamResult"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ExamResult"
                        }
                    }
                }
            }
        },
        "/exam-results/student/{student_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "成绩管理"
                ],
                "summary": "获取学生成绩",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "学生ID",
                        "name": "student_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "course_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ExamResult"
                            }
                        }
                    }
                }
            }
        },
        "/exam-results/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "成绩管理"
                ],
                "summary": "更新成绩记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "成绩ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "成绩信息",
                        "name": "result",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ExamResult"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ExamResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "成绩管理"
                ],
                "summary": "删除成绩记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "成绩ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Course": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ExamResult": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "course": {
                    "$ref": "#/definitions/models.Course"
                },
                "course_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "exam_date": {
                    "type": "string"
                },
                "exam_name": {
                    "description": "考试名称，如\"第一次月考\"",
                    "type": "string"
                },
                "exam_type": {
                    "description": "midterm期中, final期末, quiz小测",
                    "type": "string"
                },
                "full_score": {
                    "description": "满分",
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "score": {
                    "type": "number"
                },
                "student": {
                    "$ref": "#/definitions/models.Student"
                },
                "student_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.Student": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "grade": {
                    "description": "e.g., \"初二\"",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "description": "备注",
                    "type": "string"
                },
                "parent_phone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}